fixed ocarina shader floatness

+++ data/shaders/ocarina_warp.frag.glsl
@@ -24,9 +24,9 @@ uniform float magnitude; // Magnitude of the waving effect ~ [0.01 -> 0.1]
 uniform bool separated; // Set this to true to enable image `scanlines` duplication
 
 void main() {
-    vec2 offset = vec2(magnitude*sin(sol_time*0.004+gl_FragCoord.y*2*magnitude),0);
+    vec2 offset = vec2(magnitude*sin((float)sol_time*0.004+gl_FragCoord.y*2.0*magnitude),0.0);
     if(mod(gl_FragCoord.y,2.0) > 0.5 && separated) {
-      offset *= vec2(-1,1);
+      offset *= vec2(-1.0,1.0);
     }
     vec4 tex_color = COMPAT_TEXTURE(sol_texture, sol_vtex_coord+offset);
     FragColor = tex_color * sol_vcolor;
